Change GetInPartyMenu to u32

+++ src/pokemon_storage_system.c
@@ -10250,7 +10250,7 @@ void UpdateSpeciesSpritePSS(struct BoxPokemon *boxMon)
     sJustOpenedBag = FALSE;
 }
 
-u8 GetInPartyMenu(void)
+u32 GetInPartyMenu(void)
 {
     return sInPartyMenu;
 }
